requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
google-generativeai
python-dotenv
